IndexedDB: More WPT tests for reading from autoincrement stores.
Auto-increment keys represent a challenge in implementations where a
central browser process owns the databases' metadata, and multiple
renderer processes can access the same database.

Specifically, IndexedDB allows application code to start a transaction
and queue requests synchronously. So, there is no opportunity for the
browser process to pass the current autoincrement key to a renderer
process before the renderer processes requests. This situation can be
handled by queueing requests, which is complex, or by crazy hacks, which
is tempting and thus quite popular.

The tests here attempt to exercise all the code paths in the crazy
hacks that work reasonably well today. The tests do not cover compound
indexes that include the primary (autoincrement) key, because that is
not handled well in any browser.

Change-Id: Ibbe38fd173d0821d329cafed449be359e2b33f6e