[css-properties-values-api] Proper computed value for <color>.
As of an imminent spec change, <color> values no longer compute to their
specified value, but behave like <color> values in general.

Note that color keywords are parsed as CSSIdentifierValue (holding a
CSSValueID), and custom idents are parsed as CSSCustomIdentValue (holding a
String), which is why the "tomato | <color>" works as it should.

R=futhark@chromium.org

Bug: 641877
Change-Id: I946536a9d54dbaa7af589cb99acdba72f37fe016